refactor: restructure project
Using the workspace feature to have multiple libraries. `chess` and `engine` libs for now along with the multiple bins that we already had. Also refactors all the bins to be standalone files where possible (except byte-knight).
