+++ chess/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "chess"
+version = "0.1.0"
+# 2024 isn't stable yet
+edition = "2021"
+
+[lib]
+name = "chess"
+test = true
+
+[dependencies]
+anyhow = "1.0.93"
+arrayvec = "0.7.6"
+clap = { version = "4.5.21", features = ["derive"] }
+colored = "2.1.0"
+console = { version = "0.15.8", features = ["windows-console-colors"] }
+csv = "1.3.1"
+indicatif = { version = "0.17.9", features = ["rayon"] }
+itertools = "0.13.0"
+rand = { version = "0.8.5", features = ["small_rng"] }
+rand_chacha = "0.3.1"
+rayon = "1.10.0"
+serde = { version = "1.0.215", features = ["derive"] }
+stopwatch = "0.0.7"
+thiserror = "2.0.3"
