+++ chess/src/color.rs
