+++ chess/src/definitions.rs
