+++ chess/src/magics.rs
