+++ chess/src/rank.rs
