+++ engine/src/lib.rs
@@ -1,0 +1,9 @@
+pub mod defs;
+pub mod engine;
+pub mod evaluation;
+pub mod input_handler;
+pub mod psqt;
+pub mod score;
+pub mod search;
+pub mod search_thread;
+pub mod tt_table;
