+++ src/bin/generate_magics.rs
