+++ src/bin/perft.rs
